requirements.txt
fastapi
uvicorn
python-dotenv
pandas
prophet
plotly
sqlalchemy
pymysql
tabulate
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-ollama
langgraph
langchain-text-splitters
langchain-huggingface
tavily-python
faiss-cpu
pypdf
sentence-transformers
flwr[simulation]
torch
transformers
peft
bitsandbytes
datasets
accelerate
scipy